LayoutTests/media/element-containing-pip-video-going-into-fullscreen-expected.txt

Tests pip video going into fullscreen should exit pip

Video going into Picture-in-Picture
EXPECTED (internals.isChangingPresentationMode(video) == 'false') OK
EXPECTED (video.webkitPresentationMode == 'picture-in-picture') OK
Element containing picture-in-picture video going into Full Screen
EXPECTED (internals.isChangingPresentationMode(video) == 'false') OK
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLDivElement]') OK
EXPECTED (video.webkitPresentationMode == 'inline') OK
END OF TEST
